Organized repository a bit.
Grammer and precedence table are now in docs directory.
Main now doesn't contain anything, it will be used for implementation of parser.
Scanner will be tested using file sc_test.c, which can be compiled using `make test`. Tests for scanner to be completed.

Unfortunately because of several changes in scanner I am forced tu merge branched master and parser.
Because current implementation of parser on parser branch needs to be rewritten I decided to delete branch parser and reimplement it on master.
It will be easier than merging branches.
Everything important will be saved for future use.